Add files via upload
I split the sound to multiple chunks. and then want to test how the ML model works. For that, I trained the chunks data on ML models, then got the labels for each label.
According to the frequency of chunk labels from full sound, I gave to that full sound audio classification label. For instance, I may have 10 chunks corresponding allergies class. But 6 chunks were classified as  allergies, and 4 were classified as pneumonia. The label for the full sound will be allergies. So the initial label is allergies, and predicted label is allergies. That's why our model classified the chunks correctly.